hindi_gec/wikiedits/requirements.txt
nltk
tqdm
lxml
more_itertools
python_Levenshtein
PyYAML